Move complain_loudly_on_segfault to cpp/segfault_handler.h
